[add]API TEST with outh

+++ server/app/Http/Controllers/ContactsController.php
@@ -7,21 +7,41 @@ use Illuminate\Http\Request;
 
 class ContactsController extends Controller
 {
+
+    public function index(){
+        return request()->user()->contacts;
+    }
+
     public function store(){
-        Contact::create($this->valiedateData());
+        request()->user()->contacts()->create($this->valiedateData());
     }
 
     public function show(Contact $contact){
+
+        if(request()->user()->isNot($contact->user)){
+            return response([],403);
+        }
+
         return $contact;
     }
 
     public function update(Contact $contact){
+
+        if(request()->user()->isNot($contact->user)){
+            return response([],403);
+        }
+
         $contact->update(
             $this->valiedateData()
         );
     }
 
     public function destroy(Contact $contact){
+
+        if(request()->user()->isNot($contact->user)){
+            return response([],403);
+        }
+
         $contact->delete();
     }
